bug fix: support env / status without pip (#992)
